TEST CODE: Running networking on main thread
** :warning: This code causes an error. We are running from the main
thread synchronously which is a baaad idea but, we'll fix that in the
next commit. :warning: **

Created TAG for logging.
Created test latitude and longitude variables for Alcatraz.
Set forecastUrl.
** :warning: NOTE: apiKey has been removed from this GitHub commit - it
should be manually added. :warning: **

Removed menu code - we're not using menus.

Created new OkHttpClient.
Created request using OkHttp's Request.Builder with our url. We then
build the request.
Created client call for our request.
Will then try to execute this call and log it verbosely.
If call fails, catch the exception and log that in error.