Set up Alert Dialog
New AlertDialogFragment class - extends DialogFragment.

In onCreateDialog(), will getActivity() and use for context. Then, we'll
use the AlertDialog.Builder builder to create a new AlertDialog, setting
our title, message and PositiveButton (With no onClickListener()).

Will then create the AlertDialog from the builder and return it.

In MainActivity, moved log statement so that it is called whether the
response is successful or not.
Created new alertUserAboutError() method that will run in the else
statement, if the response was not successful. This will construct the
AlertDialogFragment object which we can use to show the dialog with
dialog.show(), passing in Android's FragmentManager and a key of our
choice ("error_dialog" in this case).

Added the appropriate strings to strings.xml.