Checking network availability
Code looks a bit messy here but that's because it was realigned with our
new if statement.

Added permission ACCESS_NETWORK_STATE so that we can check network
availability before proceeding to gather data from the web.

Added if statement that will only run any OkHttp code if
isNetworkAvailable() returns true.

isNetworkAvailable() will get the ConnectivityManager, get the
NetworkInfo from that, initialize isAvailable as false and only set
isAvailable to true if networkInfo is not null and isConnected.
Will then return isAvailable (Boolean).

In Strings.xml, added our "Network is unavailable" message which is used
in a Toast if the if statement fails (isNetworkAvailable() returned
false).