Started work with LinearLayouts
Set up left side of LinearLayout.

We have 1 main LinearLayout with 2 children - the children have an equal
split of 50/50 with the weight sum being 100.

On the left side, our LinearLayout holds the humidityLabel and
humidityValue TextView's. These are all set up.

On the right side, we've got two placeholder TextView's. We'll give
these values and update their properties accordingly shortly.