Added refresh button
In activity_main.xml, created an ImageView for the Refresh Button and a
ProgressBar over the ImageView (Hidden until called).

In MainActivity, injected the new views.
Refactored the getting of the forecast from onCreate() to new
getForecast() method that takes in a latitude and longitude.
Set OnClickListener() so that when the ImageView is tapped on, we will
get the forecast.

Created new toggleRefresh() method that will toggle the visibility of
the refreshImageView and progressBar.
Will call toggleRefresh() when we start getting data and finish getting
data.