Refactored code to get current forecast
Created getters/setters for the Forecast class.

Instead of a Current object, we now have a Forecast object.
We'll set this to the result of parseForecastDetails() which will return
us a Forecast object with the Current weather filled in.

Updated updateDisplay() to get our Current object from the Forecast
object and refactored this so that instead of using a member variable,
we just use the current object we created in the scope of the method.