Getting daily forecast
getDailyForecast() throws JSONException.

Will create new forecast JSONObject and get the timezone.

Will then create daily JSONObject and get the JSONArray from it.

Will then create a Day[] array of the length of the data array.

For each JSONArray, create a new Day, get the JSONObject for that day,
set the properties of the Day object and put it in the days array.

After the for loop , return the days array.