Added LinearLayout with two buttons
We're going to be making activities for the Hourly forecast and the 7
Day forecast so, I've gone ahead and updated our layout to include a
LinearLayout with these two buttons in.
Both buttons share equal weight, have heights that wrap_content and
widths that match_parent (Required for LinearLayout).
Text is capitalized to follow convention (It looks better).
ID's have been set accordingly.
Background is semi-transparent.
textColor is solid white.