Using new background gradient
Created new shape drawable (Using rectangle).

The child of this is the gradient attribute - with the startColor as a
yellow and endColor as a red.
Then set the angle to 270 so it's yellow on the top vertical and red on
the bottom vertical.
Then hardcoded type as linear.

In activity_main.xml, used the new background and also added a left
margin to our 2nd button so that they don't join together.