New DailyForecastActivity
Extends ListActivity.

Updated activity_daily_forecast.xml to have a ListView with the ID
@android:id/list for the ListActivity. The activity will also use our
gradient background.
Also added a TextView with the ID @android:id/empty that will be
displayed if we have no data.
TextView connects to new String resource, no_daily_forecast_data ("There
is no data to display!"), is in the center of the screen and uses the
Android system white color.