Started work on our custom adapter
Created our two member variables:

mContext and mDays (Our array of days).

Created a constructor that will set the member variables.

Set up the methods that we must override from extending the BaseAdapter
class.
getCount() returns the length of mDays.
getItem() returns the item of mDays passed into the method.
getItemId() isn't going to be used but can be used for tagging. We must
still include the method to avoid any errors, however.

Created ViewHolder class, nested inside of the DayAdapter class.
This will hold all of our views that we're using. They will also be
public to anything that can access the class (Only DayAdapter) so that
we can reference them later.

In getView(), we will create our ViewHolder and call it holder (Which is
initialized later).
We'll then check to see if a convertView exists (The view object we want
to re-use). The first time getView() is called, it will not so, we will
need to create everything. If getView() has already been called,
however, we can recycle all of the views that we just created in the
ViewHolder.

So, what we're doing is we're setting the views the first time and,
after that, we can just access the same views from ViewHolder, without
the need to create the views again. All we need to do if they exist is
reset the data.

We also set our convertView to a LayoutInflater from the context. What
this does is it takes XML layouts and changes them into views in code
that we can use.

After setting all of the views, we will then set a tag for the
convertView which will allow us to hold all of the views together. We
can then get these views back later which we do in the else block.

Here, we set holder to the ViewHolder returned by convertView.getTag() -
this will have all of the views we stored in the tag and will give them
back to us, saving the need to re-create them.