Setting properties in convertView
Will now get the Day from the position and then set the appropriate
properties based on the Day.
Will return the views in the convertView as the holder tag.

In DailyForecastActivity, we're now setting the adapter. We still need
to get our Day[] array with our data so, we'll do that next.