Using Parcelable data
To get our data across from Day to DailyForecastActivity, we're using
Parcelable data

So, in Day.java, we're now implementing Parcelable. The
describeContents() method isn't being used.
In writeToParcel(), we're writing all of the variables to the Parcel
destination.
In the private Day constructor, we take in a parcel and deserialize it
so that it can be read. We then set the values in the same order
(Required in Parcels).
Also created a new Day constructor that does nothing just so that we can
use Day elsewhere.
Then created our CREATOR variable. This will create a new Day object,
passing in the Parcel we want to read.
The newArray() method will create a new Day array of the size passed in.

Will then pass this in to our Intent as an extra in MainActivity, using
DAILY_FORECAST as the key.

In DailyForecastActivity, will get the parcelables from the Intent and
then create a copy of this array called mDays (So that we can use it in
our adapter). Will then set the adapter.

Now, our 7 day forecast is displayed! Woohoo! Things aren't aligned
quite right though so, we'll work on that next.