Made Hour Parcelable
Hour class implements Parcelable.

describeContents() is not used.
writeToParcel() will write the current Hour variables into a dest
parcel.
Created two constructors - one that takes in a Parcel and sets the Hour
variables and another that is just a blank constructor.
CREATOR is public and will allow the creation of a new Hour object using
the Parcel passed in and a new Hour array of the size passed in.