Created hourly_list_item layout
Layout contains a RelativeLayout with 8dp padding top and bottom.
Added a tools background so that I can see the white text.

timeLabel TextView is aligned to the left of the parent and centered
vertically. TextView is white and 24sp. Added tools value of 12 PM.

iconImageView ImageView is aligned to the right of timeLabel. It is also
centered vertically. By default, it will have the src of partly_cloudy.

temperatureLabel TextView is positioned on the very right of the screen.
It will also be centered vertically, with a text color of white and
24sp. It will use the tools text of 100.

summaryLabel TextView will be positioned to the right of iconImageView
and will extend to the left of temperatureLabel (Prevents overlap). It
will also be centered vertically, with the text color of white. It will
also have left and right padding at 15dp and the tools text of Partly
cloudy.