Started work on HourAdapter
The HourAdapter extends the generic RecyclerView.Adapter class. We are
now required to use a ViewHolder and this is passed in as the generic
type.

We create the ViewHolder within the class itself (Subclass). In a
RecyclerView, the ViewHolder is also responsible for mapping the views
so in the constructor, we'll also find the views. We'll set their values
later.

onCreateViewHolder(), onBindViewHolder() and getItemCount() are all
implemented methods of RecyclerView.Adapter that we'll use later.