Binding data in HourAdapter
Updated Hour to bind data - getTemperature() will return an int rather
than a double - the double value for mTemperature will be rounded.
getIconId() will return the icon id from the Forecast class.
getHour() will return the hour as a String. It will be formatted as
"HOUR AM/PM".
We'll again pass in the time zone to get the right hour based on what is
returned from Forecast.io.
We'll then create a new Date object, passing in our time in ms (Which
we'll convert to seconds).
We'll then return the formatted time.

In HourAdapter, created a new bindHour() method in the HourViewHolder.
This will take in the Hour object we are binding then get and set the
appropriate values.