Finished adapter methods
First off, created the HourAdapter constructor where we'll get hour
Hour[] to bind the data with.

In onCreateViewHolder(), we'll create a new view from our
hourly_list_item using the LayoutInflater. We'll get the context from
the ViewGroup parent.
We can then create a new HourViewHolder using this view and return it.

In onBindViewHolder(), we'll run bindHour() off the holder passed in,
getting the Hour from the position that is also passed in.

In getItemCount(), we'll just return the length of our Hour array.