Set the adapter - Hourly Forecast finished!
Using ButterKnife to inject the RecyclerView.

Will then create a new HourAdapter and set it as the adapter for the
RecyclerView.

Also using a LayoutManager for the RecyclerView - this deals with the
positioning and recycling of views.
We tell it that this layout is going to be portrait
(LinearLayoutManager) and that it has a fixed size (For improved
performance).