Set onClickListener() for RecyclerView
Setting an onClickListener() is slightly different for a RecyclerView.
Instead, our ViewHolder itself should implement View.OnClickListener()
where the onClick() method gets implemented and the OnClickListener() is
set in the constructor (It's set to this as the class itself is the
onClickListener())
Note that we also need a context to do this so, HourAdapter will take in
the context and set it as a private member variable.
HourlyForecastActivity has to reflect this change to the constructor.